Implement new main page
Renders title, button
